requirements.txt
openai
chainlit
python-dotenv
huggingface-hub
pandas